refactor: move MaskThemeProvider
